Properly fix activity filter across versions
