+++ plugin-flex-ts-template-v2/src/feature-library/activity-skill-filter/custom-components/activity-wrapper/index.ts
@@ -1,3 +1,3 @@
-import ActivityWrapperComponent from './ActivityWrapperComponent';
+import ActivityWrapper from './ActivityWrapper';
 
-export default ActivityWrapperComponent;
+export default ActivityWrapper;
